Terminal befehle in eigene File ausgelagert
